Use Notary API directly for platform freedom (#5)
To allow this utility to be used on any operating system, dependencies on
`notarytool` for macOS or `iTMS Transporter` on linux have been removed
and replaced with direct API calls to the Apple Notary API.

As part of this change support for stapling disk images (`.dmg` files)
has been introduced removing any dependency on `notarytool` for any file
type allowing for complete freedom.

Ground work has also been laid to allow for code signing files using this
utility as well, which would completely remove the need on macOS for
pre-distribution tasks.