Fix CodeDirectory decode versioning, remove keychain testing code
Fixed a bug in the blob decoder for `CodeDirectory`, it was taking the
version once before the decode loop which was resulting only the base to
be decoded and none of the support versions which would result in broken
code signature in dmgs on stapling.

Also removed proof of concept code that was used to work out how to use
the keychain APIs, it was no longer needed and shouldn't of been
included.